Create web config class
Create web config class to configure web mvc in a new Class
